fix(database): prevent partial repo updates from clobbering omitted fields
Repository update() methods that merge a partial patch with a plain
`{ ...current, ...patch }` spread can clobber omitted fields: JS spread
copies keys whose value is undefined, so any caller that forwards an
omitted field as undefined overwrites the stored value.

PR #27 already fixed ShoppingListItemRepo.update via an inline
mergeDefinedPatch helper: its PATCH route forwards { quantity, note,
checked }, so a note-only or Shop-mode { checked } update was an
active data-loss bug. CategoryRepo.update and ShoppingListRepo.update
still use the raw spread; their routes happen to sanitize the patch
today, so there the clobber is latent rather than active.

Promote the helper into a shared database/merge-patch.ts module and
route ShoppingListItemRepo, CategoryRepo, and ShoppingListRepo
update() through it, so the guarantee lives in one tested place rather
than being copied per repo. Only patch keys whose value is not
undefined are applied; defined falsy values (false, 0, "") still apply.

Also fold PR #27's inline-helper tests (shopping-list-item.repo.test.ts)
into merge-patch.test.ts, expanded with empty-patch and no-mutation
cases. Plain `deno test` runs without --unstable-kv, so these KV-free
tests of the pure helper never open Deno KV.
