+++ database/merge-patch.test.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from "jsr:@std/assert@^1.0.19";
-import { mergeDefinedPatch } from "@/database/shopping-list-item.repo.ts";
+import { mergeDefinedPatch } from "@/database/merge-patch.ts";
 
 /** Minimal shape covering the fields exercised below. Typed explicitly (rather
  *  than inferred from each literal) so patches touching a key `current` didn't
@@ -33,4 +33,17 @@ Deno.test("mergeDefinedPatch — defined falsy values still apply", () => {
   const current: Partial<TestItem> = { checked: true };
   const result = mergeDefinedPatch(current, { checked: false });
   assertEquals(result, { checked: false });
+});
+
+Deno.test("mergeDefinedPatch — an empty patch returns an unchanged copy", () => {
+  const current: Partial<TestItem> = { quantity: 3, note: "keep" };
+  const result = mergeDefinedPatch(current, {});
+  assertEquals(result, { quantity: 3, note: "keep" });
+});
+
+Deno.test("mergeDefinedPatch — does not mutate the original current object", () => {
+  const current: Partial<TestItem> = { quantity: 1, note: "orig" };
+  const result = mergeDefinedPatch(current, { note: "changed" });
+  assertEquals(current, { quantity: 1, note: "orig" });
+  assertEquals(result, { quantity: 1, note: "changed" });
 });
